\name{radiata}
\docType{data}
\alias{radiata}

\title{The Pinus Radiata dataset}
\description{
  Pinus radiata is one of the mostly widely planted species in Chile and is planted in a wide
  array of soil types and regional climates. The plots were located in the Escuadron sector,
  south of Concepcion, in the southern portion of Chile and has an area of 1244.43 hectares.
}
\usage{data(radiata)}
\format{
  A data frame with 468 observations on the following 6 variables.
  \describe{
    \item{xpos}{ \code{x}-coordinates.}
    \item{ypos}{ \code{y}-coordinates.}
    \item{basal}{ basal area measurements.}
    \item{height}{ dominant tree height.}
    \item{altitude}{ altitude in meters.}
    \item{slope}{ slope of the terrain plot.}
  }
}
\source{
  Cuevas, F., Porcu, E., Vallejos, R. (2013).
  Study of spatial relationships between two sets of variables: A nonparametric approach.
  \emph{Journal of Nonparametric Statistics} \bold{25}, 695--714.

  Vallejos, R., Osorio, F., Bevilacqua, M. (2020+).
  \emph{Spatial Relationships Between Two Georeferenced Variables: with Applications in R}.
  Springer, New York.
}
\keyword{datasets}
