\name{texmos2}
\docType{data}
\alias{texmos2}

\title{USC texture mosaic number 2}
\description{
  Gray-level texture map with information about mosaic composed of eight different
  texture samples taken from the Brodatz texture book, available from USC-SIPI image
  database.
}
\usage{data(texmos2)}
\format{
  A grayscale matrix of size 512-by-512.
}
\source{
  Brodatz, P. (1966).
  \emph{Textures: A Photographic Album for Artist and Designers}.
  Dover Publications, New York.

  USC-SIPI image database, URL: \url{http://sipi.usc.edu/database/}
}
\keyword{datasets}
