\name{wheat}
\docType{data}
\alias{wheat}

\title{Brodatz texture image, Straw (D15)}
\description{
  Gray-level texture image depicting wheat taken from the Brodatz texture book,
  available from USC-SIPI image database.
}
\usage{data(wheat)}
\format{
  A grayscale matrix of size 512-by-512.
}
\source{
  Brodatz, P. (1966).
  \emph{Textures: A Photographic Album for Artist and Designers}.
  Dover Publications, New York.
}
\keyword{datasets}
